requirements.txt
aioredis==1.2.0
amqp==2.5.0
asn1crypto==0.24.0
astroid==2.3.1
async-timeout==2.0.1
attrs==19.3.0
autobahn==19.6.2
Automat==0.7.0
backports.csv==1.0.7
base58check==1.0.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
coinaddr==1.0.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==3.2
django-4-jet==1.0.8
django-autoslug==1.9.8
django-crequest==2018.5.11
django-environ==0.4.5
djangorestframework==3.11.0
dnspython==1.16.0
email-validator==1.0.4
et-xmlfile==1.0.1
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
jdcal==1.4.1
kombu==4.6.3
lazy-object-proxy==1.4.2
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
pep8==1.7.1
Pillow==7.1.0
#pkg-resources==0.0.0
#psycopg2==2.8.3
#psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pysha3==1.0.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
qrcode==6.1
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
Twisted==20.3.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wrapt==1.11.2
xlrd==1.2.0
django-ckeditor==6.0.0
django-embed-video==1.3.3
